requirements-gpu.txt
numpy==1.16.4
imageio
tensorflow-gpu==1.8.0
keras==2.2.5
h5py
graphviz
pydot
matplotlib
ipykernel
tqdm
streamlit
pytest
